refactored the code
I think that essential change that solved #107 was moving the setLoggedOn later into the promise chain
